% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author_as_df.R
\name{author_as_df}
\alias{author_as_df}
\title{Make a data frame with the information of an author from a publication retrieved from Entrez}
\usage{
author_as_df(author)
}
\arguments{
\item{author}{A XML node with the information of an author from a paper}
}
\value{
A dataframe with the first, last and full name of the author, and their affiliation
}
\description{
Make a data frame with the information of an author from a publication retrieved from Entrez
}
\examples{
pmids <- get_pmids("Benjamin Valderrama")

list_of_author_pubs <- lapply(pmids, get_authors_node) |>
lapply(function(x){lapply(x, author_as_df) } )
}
