% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author_gs_profile.R
\name{author_gs_profile}
\alias{author_gs_profile}
\title{Get the Google Scholar profile of an author}
\usage{
author_gs_profile(search_term)
}
\arguments{
\item{search_term}{A character vector in the same shape as queries used in entrez with the name of an author}
}
\value{
a list with information retrieved from the authors Google scholar profile
}
\description{
Get the Google Scholar profile of an author
}
\examples{
author_gs_profile("Benjamin Valderrama[AU]")
}
