% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_author_name.R
\name{clean_author_name}
\alias{clean_author_name}
\title{Get first and last names from a entrez's search term}
\usage{
clean_author_name(search_term)
}
\arguments{
\item{search_term}{A character vector in the same shape as queries used in entrez with the name of an author}
}
\value{
A character vector with just the first and last name of the author in the search term
}
\description{
Get first and last names from a entrez's search term
}
\examples{
clean_author_name("Benjamin Valderrama[AU]")
}
