% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_authors_node.R
\name{get_authors_node}
\alias{get_authors_node}
\title{Get a nodes with all authors in a publication}
\usage{
get_authors_node(pmid)
}
\arguments{
\item{pmid}{A character vector with a PubMed ID of a publication made by the author}
}
\value{
A list of nodes of each author in a publication
}
\description{
Get a nodes with all authors in a publication
}
\examples{
pmids <- get_pmids("Benjamin Valderrama")
pmid <- pmids[1]
get_authors_node(pmid)
}
