% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coauthors.R
\name{get_coauthors}
\alias{get_coauthors}
\title{Get coauthors of an author of interest}
\usage{
get_coauthors(search_term)
}
\arguments{
\item{search_term}{A character vector in the same shape as queries used in entrez with the name of an author}
}
\value{
a data frame with the information of the coauthors of each paper published by the author of interest
}
\description{
Get coauthors of an author of interest
}
\examples{
get_coauthors("Benjamin Valderrama[AU]")
}
