% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pmids.R
\name{get_pmids}
\alias{get_pmids}
\title{Get all PMID's of an author of interest}
\usage{
get_pmids(search_term)
}
\arguments{
\item{search_term}{}
}
\value{
A list with all the PMID's associated to the author available in pubmed
}
\description{
Get all PMID's of an author of interest
}
\examples{
get_pmids("Benjamin Valderrama[AU]")
get_pmids("Benjamin Valderrama[AU] microbiome")
}
