% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pmid_to_author_df.R
\name{add_pmid_to_author_df}
\alias{add_pmid_to_author_df}
\title{Adds PubMed ID to the dataframe of an author generated by \code{author_as_df}}
\usage{
add_pmid_to_author_df(author_df, pmid)
}
\arguments{
\item{author_df}{a dataframe of an author made by \code{author_as_df()}}

\item{pmid}{A character vector with a PubMed ID of a publication made by the author}
}
\value{
the author_df given as a parameter with an extra column with the PubMed ID given in pmid
}
\description{
Adds PubMed ID to the dataframe of an author generated by \code{author_as_df}
}
\examples{
pmids <- get_pmids("Benjamin Valderrama")

list_of_author_pubs <- lapply(pmids, get_authors_node) |>
lapply(function(x){lapply(x, author_as_df) } ) |>
lapply(function(x) do.call(rbind, x))

coauthors_df <- base::Map(f = add_pmid_to_author_df, list_of_author_pubs, as.list(pmids)) |>
do.call(what = "rbind")
}
